--- src/components/Burger/Burger.jsx
@@ -1,10 +1,0 @@
-import { createPortal } from 'react-dom';
-import MobileMenu from './MobileMenu.jsx/MobileMenu';
-
-const Burger = () => {
-  const burgerRoot = document.getElementById('burger-root');
-
-  return createPortal(<MobileMenu />, burgerRoot);
-};
-
-export default Burger;
