--- src/components/Burger/MobileMenu.jsx/index.js
@@ -1,1 +1,0 @@
-export { default } from './MobileMenu';
